requirements.txt
# API
fastapi
uvicorn[standard]

# Networking
requests
python-hosts

# Provisioning
pypxe

# Utils
GitPython
python-apt
apscheduler